+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Config Curator ChangeLog
 
+## HEAD
+
+- PackageLookup#tools separates tool identifier and tool command.
+- PackageLookup#installed? checks for tool command before lookup.
+
 ## 0.0.1
 
 - Initial development release.